:boom: Allow trailing spaces before new lines
When using Lipgloss to align text to the right, trailing spaces are add
as padding.

Reflow is removing this padding (by design) which is causing alignment
issues. Each line is pushed against the right border. This only occurs
when using no border, once a border is added this problem does not
happen.

Tests are failing now due to this change however this solves a specific
use case.